Consumiendo api de otro servicio

+++ product-service/src/main/java/com/tecsup/app/micro/product/entity/ProductEntity.java
@@ -23,5 +23,5 @@ public class ProductEntity {
     private BigDecimal price;
     private Integer stock;
     private String category;
-    private Integer createdBy;
+    private Long createdBy;
 }
